requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
Django==4.1.13
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.7
laftel==1.3.1
Markdown==3.6
multidict==6.0.5
mysqlclient==2.2.0
pytz==2024.1
sqlparse==0.2.4
yarl==1.9.4
